remove array package and add slice package.
