+++ .github/workflows/test.yml
@@ -38,5 +38,4 @@ jobs:
       - name: Send coverage
         env:
           COVERALLS_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          COVERALLS_PARALLEL: true
         run: goveralls -coverprofile=covprofile -service=github
