+++ slice/seat_test.go
@@ -1,8 +1,8 @@
-package array_test
+package slice_test
 
 import (
 	"fmt"
-	"github.com/jtyoui/gotool/array"
+	"github.com/jtyoui/gotool/slice"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -24,18 +24,18 @@ func TestReverse(t *testing.T) {
 
 	for i := 0; i < len(value); i++ {
 		data := value[i]
-		array.Reverse(data)
+		slice.Reverse(data)
 		assert.Equal(t, predict[i], data)
 	}
 }
 
 func ExampleReverse() {
 	data := []string{"a", "b", "c"}
-	array.Reverse(data)
+	slice.Reverse(data)
 	fmt.Println(data)
 
 	value := []int{1, 2, 3}
-	array.Reverse(value)
+	slice.Reverse(value)
 	fmt.Println(value)
 	// Output:
 	// [c b a]
@@ -44,10 +44,16 @@ func ExampleReverse() {
 
 func ExampleCopyReverse() {
 	data := []string{"a", "b", "c"}
-	value := array.CopyReverse(data)
+	value := slice.CopyReverse(data)
 	fmt.Println(data)
 	fmt.Println(value)
 	// Output:
 	// [a b c]
 	// [c b a]
+}
+
+func TestRemoveElement(t *testing.T) {
+	value := []int{1, 2, 3, 4}
+	remove := slice.RemoveElement(value, 1)
+	assert.Equal(t, remove, []int{2, 3, 4})
 }
